Rename autogen/ docs references to ag2/
