+++ docs/docs/en/api/fastagency/runtimes/ag2/agents/whatsapp/WhatsAppAgent.md
@@ -1,0 +1,11 @@
+---
+# 0.5 - API
+# 2 - Release
+# 3 - Contributing
+# 5 - Template Page
+# 10 - Default
+search:
+  boost: 0.5
+---
+
+::: fastagency.runtimes.ag2.agents.whatsapp.WhatsAppAgent
