+++ docs/docs/en/user-guide/runtimes/index.md
@@ -2,12 +2,12 @@
 
 **FastAgency** is a flexible, open-source framework designed to accelerate the transition from prototype to production for multi-agent AI workflows. If you're a developer working with agentic frameworks like **AG2**, FastAgency provides the tools to rapidly scale your projects into fully operational applications. A key feature of FastAgency is its support for multiple runtimes, allowing developers to switch between or combine various frameworks based on project needs. With multi-framework compatibility, FastAgency ensures that your workflows are not only optimized for today’s needs but also future-proofed as new frameworks emerge.
 
-## [AG2](autogen/index.md)
+## [AG2](ag2/index.md)
 The **AG2** runtime is central to FastAgency’s architecture and provides a powerful foundation for multi-agent workflows. AG2 allows developers to define workflows in Python, leveraging large language models (LLMs) such as GPT to handle communication and collaboration between agents. It enables rapid prototyping and deployment of workflows that involve tasks like decision-making, customer service, or research.
 
 FastAgency seamlessly integrates with AG2, helping you transition from development in Jupyter notebooks to fully deployed applications. Whether you're managing conversational agents or orchestrating task automation, the AG2 runtime in FastAgency helps you streamline the path from prototype to production.
 
-For more information on using AG2 within FastAgency, visit the [AG2 section](autogen/index.md).
+For more information on using AG2 within FastAgency, visit the [AG2 section](ag2/index.md).
 
 ---
