+++ docs/docs_src/getting_started/no_auth/nats_n_fastapi/my_fastagency_app/my_fastagency_app/workflow.py
@@ -3,7 +3,7 @@ from typing import Any
 
 from autogen.agentchat import ConversableAgent
 from fastagency import UI
-from fastagency.runtimes.ag2 import AutoGenWorkflows
+from fastagency.runtimes.autogen import AutoGenWorkflows
 
 llm_config = {
     "config_list": [
