+++ tests/runtime/ag2/tools/test_whatsapp.py
@@ -3,8 +3,8 @@ from typing import Any
 import pytest
 from autogen import ConversableAgent, UserProxyAgent
 
-from fastagency.runtimes.ag2.agents.whatsapp import WhatsAppAgent
 from fastagency.runtimes.ag2.ag2 import Toolable
+from fastagency.runtimes.ag2.agents.whatsapp import WhatsAppAgent
 from fastagency.runtimes.ag2.tools import WhatsAppTool
